CHORE : LottoMatchResult 기능을 Rank로 옮기면서 생긴 변화 반영
