+++ src/main/java/lotto/controller/LottoController.java
@@ -11,17 +11,17 @@ import java.util.List;
 public class LottoController {
     public static void startLotto() {
         int money = InputView.inputMoney();
-        Lottos lottos = LottoGenerator.createLottos(LottoPurchase.purchase(money));
+        Lottos lottos = LottoGenerator.createLottos(new LottoShop(money).calculatePurchaseCount());
         ResultView.printLotto(lottos);
 
         List<Integer> winNumbers = StringUtils.parseWinNumbers(InputView.inputWinNumbers());
         int bonus = InputView.inputBonus();
 
         for (Lotto lotto : lottos.getLottos()) {
-            LottoMatchResult.putMatchResult(LottoMatch.match(lotto, new WinningLotto(winNumbers, bonus)));
+            Rank.putMatchResult(LottoMatch.match(lotto, new WinningLotto(winNumbers, bonus)));
         }
 
-        HashMap<Rank, Integer> matchResult = LottoMatchResult.getMatchResult();
+        HashMap<Rank, Integer> matchResult = Rank.getMatchResult();
         ResultView.printResult(matchResult, LottoRevenue.calculateRevenue(money, LottoRevenue.revenueTotal(matchResult)));
     }
 }
