--- src/test/java/lotto/LottoNumberTest.java
@@ -1,13 +1,0 @@
-package lotto;
-
-import lotto.domain.LottoNumber;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class LottoNumberTest {
-    @Test
-    void create() {
-        assertThat(new LottoNumber(3)).isEqualTo(new LottoNumber(3));
-    }
-}
