+++ src/test/java/lotto/domain/LottoGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import lotto.domain.Lotto;
 import lotto.domain.LottoGenerator;
