+++ src/test/java/lotto/domain/LottoMatchTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import lotto.domain.*;
 import org.junit.jupiter.api.BeforeEach;
