+++ src/test/java/lotto/domain/LottoRevenueTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto;
 
+import lotto.domain.Rank;
+import lotto.domain.LottoRevenue;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
