+++ src/test/java/lotto/domain/LottoShopTest.java
@@ -1,19 +1,23 @@
 package lotto.domain;
 
+import lotto.domain.LottoPurchase;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-public class LottoShopTest {
+public class LottoPurchaseTest {
     @Test
-    void 구매_수() {
-        assertThat(LottoShop.countPurchase(3000)).isEqualTo(3);
+    void 구매_개수() {
+        int money = 3000;
+        assertThat(new LottoPurchase(money).getPurchase()).isEqualTo(3);
     }
 
     @Test
-    void 구매금액_오류() {
+    void 구매_금액_오류() {
+        int money = 3500;
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> LottoShop.purchaseTicket(3500))
-                .withMessageMatching("구매 금액은 1000 단위로 입력해야 합니다.");
+                .isThrownBy(() -> new LottoPurchase(money))
+                .withMessageMatching("구매 금액은 1000원 단위로 입력해주세요.");
     }
 }
